Add configuration management, database initialization, and core application structure for Second Brain Database
- Implemented configuration management in `config.py` to handle environment detection and load settings from a user config file or defaults.
- Established MongoDB client and database connection in `database.py`.
- Created the main Flask application entry point in `main.py`, including route setup, middleware, and error handling.
- Initialized user module and versioned submodules for emotion tracking and notes management.
- Developed CRUD operations for emotion tracking in `emotion_tracker/model.py` and `emotion_tracker/routes.py`.
- Implemented CRUD operations for notes management in `notes/model.py` and `notes/routes.py`.
- Added decorators for user authentication and authorization in `utils/decorators/privileged.py`.
