--- src/Second_Brain_Database/__init__.py
@@ -1,8 +1,0 @@
-"""
-__init__.py
-
-Initializes the Second Brain Database package.
-
-Author: Rohan (refactored by GitHub Copilot)
-Date: 2025-06-11
-"""
