--- src/Second_Brain_Database/admin/v1/plans/__init__.py
@@ -1,8 +1,0 @@
-"""
-__init__.py
-
-Initializes the admin plans module for v1.
-
-Author: Rohan (refactored by GitHub Copilot)
-Date: 2025-06-11
-"""
