--- src/Second_Brain_Database/auth/__init__.py
@@ -1,8 +1,0 @@
-"""
-__init__.py
-
-Initializes the authentication module.
-
-Author: Rohan (refactored by GitHub Copilot)
-Date: 2025-06-11
-"""
