--- src/Second_Brain_Database/user/v1/emotion_tracker/__init__.py
@@ -1,8 +1,0 @@
-"""
-__init__.py
-
-Initializes the emotion tracker module for user v1.
-
-Author: Rohan (refactored by GitHub Copilot)
-Date: 2025-06-11
-"""
