--- src/Second_Brain_Database/user/v1/notes/__init__.py
@@ -1,9 +1,0 @@
-"""
-__init__.py
-
-Initializes the notes module for user v1.
-
-Author: Rohan (refactored by GitHub Copilot)
-Date: 2025-06-11
-"""
-# Initialize the notes module
