+++ src/second_brain_database/__init__.py
@@ -1,0 +1,8 @@
+"""
+__init__.py
+
+Initializes the Second_Brain_Database package.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
