+++ src/second_brain_database/admin/v1/plans/__init__.py
@@ -1,0 +1,8 @@
+"""
+__init__.py
+
+Initializes the admin plans module for v1.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
