+++ src/second_brain_database/assets/banner.png
