+++ src/second_brain_database/auth/__init__.py
@@ -1,0 +1,8 @@
+"""
+__init__.py
+
+Initializes the authentication module.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
