+++ src/second_brain_database/user/v1/__init__.py
@@ -1,0 +1,8 @@
+"""
+__init__.py
+
+Initializes the user v1 module.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
