+++ src/second_brain_database/user/v1/emotion_tracker/__init__.py
@@ -1,0 +1,8 @@
+"""
+__init__.py
+
+Initializes the emotion tracker module for user v1.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
