+++ src/second_brain_database/user/v1/notes/__init__.py
@@ -1,0 +1,9 @@
+"""
+__init__.py
+
+Initializes the notes module for user v1.
+
+Author: Rohan Batra
+Date: 2025-06-11
+"""
+# Initialize the notes module
